requirements.txt
django
djangorestframework
django-cors-headers 
pillow 
django-allauth 
django-environ 
django-paypal 